+++ src/star.cpp
@@ -6,8 +6,7 @@
 #include <cmath>
 #include <random>
 
-
-std::random_device rd;
+std::random_device imp_rd;
 
 //-------------------------------------------------------------------------------
 //			player, ship, and mission classes
@@ -42,135 +41,6 @@ int Player::getRepair(){
 
 int Player::getBarter(){
 	return bartering;
-}
-
-
-//-------------------------------------------------------------------------------
-//			Sun and Planets
-//-------------------------------------------------------------------------------
-Sun::Sun() {
-	sunPos = (Vector2) {
-		SCREENWIDTH/2,
-		SCREENHEIGHT/2
-	};
-
-	sunRadius = 30;
-}
-
-void Sun::DrawSun() {
-	DrawCircleV(sunPos, sunRadius, (Color) {252, 255, 252, 255});
-}
-
-Planet::Planet() {
-	std::uniform_int_distribution<int> rand_angle(0, 360);
-	std::uniform_int_distribution<int> rand_radius(3, 15);
-	std::uniform_int_distribution<int> rand_dist(50, 450);
-	std::uniform_int_distribution<int> rand_conic(0, 6);
-	std::uniform_int_distribution<int> rand_RGB(25, 255);
-
-	orbitAngle = rand_angle(rd);
-	radius = rand_radius(rd);
-	orbitDistance = rand_dist(rd);
-	conicRotation = rand_conic(rd);
-	mass = pow(radius, 2) * 100;
-	conicScale = orbitDistance / mass;
-	alpha = 0.0f;
-
-	unsigned char colors[3] = {
-		(unsigned char)(rand_RGB(rd) ),
-		(unsigned char)(rand_RGB(rd) ),
-		(unsigned char)(rand_RGB(rd) )
-	};
-	color = (Color) {colors[0], colors[1], colors[2], 255};
-	orbitColor = (Color) {colors[0], colors[1], colors[2], 55};
-
-	orbitOn = false;
-}
-
-void Planet::UpdatePlanet() {
-	orbitRadius = orbitDistance / (1 - conicScale * cos(orbitAngle - conicRotation));
-
-	pos = (Vector2) {
-		sunPos.x + cos(orbitAngle) * orbitRadius,
-		sunPos.y + sin(orbitAngle) * orbitRadius 
-	};
-	
-	for(int p=0; p<ORBITALPOINTS; p++) {
-		float distScaler = 1 / distFromMouse;
-		float splineAngleA = orbitAngle - (p * distScaler);
-		float splineAngleB = orbitAngle + (p * distScaler);
-
-		orbitPointsAhead[p] = {
-			sunPos.x + cos(splineAngleA) * (orbitDistance / (1 - conicScale * cos(splineAngleA - conicRotation)) ),
-			sunPos.y + sin(splineAngleA) * (orbitDistance / (1 - conicScale * cos(splineAngleA - conicRotation)) )
-		};
-
-		orbitPointsBehind[p] = {
-			sunPos.x + cos(splineAngleB) * (orbitDistance / (1 - conicScale * cos(splineAngleB - conicRotation)) ),
-			sunPos.y + sin(splineAngleB) * (orbitDistance / (1 - conicScale * cos(splineAngleB - conicRotation)) )
-		};
-	}
-
-	for(int p=0; p<ORBITALPOINTSFULL; p++) {
-		float splineAngleF = ((M_PI / ORBITALPOINTSFULL) * p * 2.041);
-
-		orbitPointsFull[p] = {
-			sunPos.x + cos(splineAngleF) * (orbitDistance / (1 - conicScale * cos(splineAngleF - conicRotation)) ),
-			sunPos.y + sin(splineAngleF) * (orbitDistance / (1 - conicScale * cos(splineAngleF - conicRotation)) )
-		};
-	}
-
-	distFromMouse = GetDist(pos, GetMousePosition() );
-	distFromSun = GetDist(pos, sunPos);
-
-	if (distFromMouse < 10) {
-		distFromMouse = 10;
-	}
-
-	if (orbitAngle > 0) {
-		orbitAngle -= (0.000001 / distFromSun) * mass;
-	}
-	else {
-		orbitAngle = 360;
-	}
-
-}
-
-void Planet::DrawPlanet(bool doDrawOrbital) {
-	if (((CheckCollisionPointCircle(GetMousePosition(), pos, radius)
-	|| CheckCollisionPointCircle(GetMousePosition(), sunPos, 30) )
-	&& doDrawOrbital == true)
-	|| orbitOn == true) {
-		DrawSplineLinear(orbitPointsFull, ORBITALPOINTSFULL, 2, orbitColor );
-	}
-	else if (CheckCollisionPointCircle(GetMousePosition(), pos, PLANETBOUNDS) ) {
-		DrawSplineLinear(orbitPointsAhead, ORBITALPOINTS, 2, orbitColor );
-		DrawSplineLinear(orbitPointsBehind, ORBITALPOINTS, 2, orbitColor );
-	}
-	
-	DrawCircleV(pos, radius, color);
-}
-
-void Planet::RegisterPlanetClicked() {
-	if (CheckCollisionPointCircle(GetMousePosition(), pos, radius)
-	&& IsMouseButtonPressed(MOUSE_LEFT_BUTTON) ) {
-		if (orbitOn == false) {
-			orbitOn = true;
-		}
-		else if (orbitOn == true) {
-			orbitOn = false;
-		}
-	}
-}
-
-void DrawAndUpdateSolarSystem(Sun sun, Planet *planet, bool doDrawOrbital) {
-	sun.DrawSun();
-	
-	for (int i=0; i<NUMPLANETS; i++) {
-		planet[i].UpdatePlanet();
-		planet[i].RegisterPlanetClicked();
-		planet[i].DrawPlanet(doDrawOrbital);
-	}
 }
 
 
@@ -280,7 +150,7 @@ void PTXStarAnim(PTX *ptx, float counter){
 		if (ptx[i].halflife == 0 ) {
 			std::uniform_int_distribution<int> rand_chance(0, 400);
 
-			int chance = rand_chance(rd);
+			int chance = rand_chance(imp_rd);
 			
 			if (chance == 1) {
 				std::uniform_int_distribution<int> rand_dist(0, 45);
@@ -288,9 +158,9 @@ void PTXStarAnim(PTX *ptx, float counter){
 				std::uniform_int_distribution<int> rand_screen_x(1, SCREENWIDTH);
 				std::uniform_int_distribution<int> rand_screen_y(1, SCREENHEIGHT);
 
-				ptx[i].dist += rand_dist(rd) * 0.01f;
+				ptx[i].dist += rand_dist(imp_rd) * 0.01f;
 
-				int colorChance = rand_color_selector(rd);
+				int colorChance = rand_color_selector(imp_rd);
 
 				switch (colorChance) {
 					case 0: ptx[i].color = {220, 233, 255, 255};
@@ -306,7 +176,7 @@ void PTXStarAnim(PTX *ptx, float counter){
 				}
 
 				ptx[i].halflife = 1;
-				ptx[i].pos = {float(rand_screen_x(rd) ), float(rand_screen_y(rd) )};
+				ptx[i].pos = {float(rand_screen_x(imp_rd) ), float(rand_screen_y(imp_rd) )};
 			}
 		}
